fix(build): fix issues with 'friend' functions
Move the declaration of blt8() out of the class, friend declarations cannot
have default values for parameters.

Declare kiirdesceket() as non static to fix the error that it is first declared
as non-static in TOPOL.H and then as static in TOPOL.CPP.

Add function declarations for betoltecseteket(), loadrecek() and saverecek().
These were only defined as 'friend' functions, but need proper declarations to
be used outside of the corresponding classes.